@@ -19,7 +19,7 @@ use Psr\Http\Message\ResponseInterface;
 class ApiClientTest extends UnitTestCase
 {
     /**
-     * @var ClientInterface|\PHPUnit_Framework_MockObject_MockObject
+     * @var ClientInterface|\PHPUnit\Framework\MockObject\MockObject
      */
     private $http;
 
@@ -33,7 +33,7 @@ class ApiClientTest extends UnitTestCase
      */
     private $targetUrl;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->http = $this->createMock(ClientInterface::class);
         $this->client = new ApiClient($this->http);
