@@ -15,7 +15,7 @@ class ReferenceTest extends DatabaseTestCase
     /** @var Reference */
     private $ref;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->ref = self::$db->getReference(self::$refPrefix);
     }
@@ -83,7 +83,7 @@ class ReferenceTest extends DatabaseTestCase
         $ref = $this->ref->getChild(__FUNCTION__);
         $key = $ref->push()->getKey();
 
-        $this->assertInternalType('string', $key);
+        $this->assertIsString($key);
         $this->assertSame(0, $ref->getSnapshot()->numChildren());
     }
 
