@@ -18,7 +18,7 @@ class CreateUserTest extends IntegrationTestCase
      */
     private $auth;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->auth = self::$factory->createAuth();
     }
