@@ -19,7 +19,7 @@ class UpdateUserTest extends IntegrationTestCase
      */
     private $auth;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->auth = self::$factory->createAuth();
     }
