@@ -18,7 +18,7 @@ class StorageTest extends IntegrationTestCase
      */
     private $storage;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->storage = self::$factory->createStorage();
     }
