@@ -29,7 +29,7 @@ final class GuzzleApiClientHandlerTest extends TestCase
     /** @var GuzzleApiClientHandler */
     private $handler;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->client = $this->prophesize(ClientInterface::class);
 
