@@ -25,6 +25,6 @@ class LimitToFirstTest extends UnitTestCase
     {
         $filter = new LimitToFirst(3);
 
-        $this->assertContains('limitToFirst=3', (string) $filter->modifyUri(new Uri('http://domain.tld')));
+        $this->assertStringContainsString('limitToFirst=3', (string) $filter->modifyUri(new Uri('http://domain.tld')));
     }
 }
