@@ -17,6 +17,6 @@ class ShallowTest extends UnitTestCase
     {
         $filter = new Shallow();
 
-        $this->assertContains('shallow=true', (string) $filter->modifyUri(new Uri('http://domain.tld')));
+        $this->assertStringContainsString('shallow=true', (string) $filter->modifyUri(new Uri('http://domain.tld')));
     }
 }
